Matrix Alignment and Validation Added
In the `Encoder` class, the `EncodeFrame` method has been updated to align the `curr` and `pred` matrices using `Transform.AlignToBlock`. In the `Transform` class, the `Sub` method has been reworked to validate matrix sizes and throw an `ArgumentException` if there are discrepancies. A new `AlignToBlock` method has been added to adjust matrices to the block size `bs`.